Better baseline mat_mul schedule
It's still bad, because we should be staging the inputs and factoring
the reduction, but this is more respectable than the schedule that was
there.

Large reductions should be wrapped in another function so that the
reduction can happen into registers instead of into global memory. The
original code had lots of stores in the inner loop, where there should
only be loads in a mat mul.